requirements.txt
absl-py==0.3.0
astor==0.7.1
autopep8==1.3.5
awsebcli==3.14.3
botocore==1.10.65
cement==2.8.2
certifi==2018.4.16
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.15.0
pathspec==0.5.5
Pillow==5.2.0
protobuf==3.6.0
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
Werkzeug==0.14.1
wincertstore==0.2
